more command system stuff, still trying 2 fix opacity lul.

+++ src/main/java/me/srgantmoomoo/postman/client/setting/settings/ColorSetting.java
@@ -39,8 +39,8 @@ public class ColorSetting extends Setting implements com.lukflug.panelstudio.set
 		return this.value.getRGB()&0xFFFFFFFF+(rainbow?1:0)*0x100000000L;
 	}
 	
-	public void fromInteger (int number) {
-		this.value = new JColor(number&0xFFFFFFFF,true);
+	public void fromInteger (long number) {
+		this.value = new JColor((int) (number&0xFFFFFFFF),true);
 		this.rainbow = ((number&0x100000000L)!=0);
 	}
